requirements.txt
requests>=2.28.0
tqdm
pyfiglet
termcolor
dnspython==2.2.1
urllib3>=2.0.0
chardet>=4.0.0
aiohttp>=3.7.4
rich>=10.0.0
